Found my hardfault problem. It was actually caused by a previous stackoverflow error, which was triggering my panic(), but my panic() didn't disable chibios, so after the assert failed, other threads would continue running, causing no good.
I increased the stack size for everything for the moment, and I'll reduce as needed.

I added the debug files to try to sort debug stuff out, so that helps sometimes.

I've started a file-checker program, written in Python, that helps determine if my .KBB files are correct. I'm currently getting 10% dropped messages, which is expected, since my processing is dumb.