Added Baud rate to config. Added FeatureA to config. Fixed polarity on the gpio get/set function for button & serial power. Logger now exits properly - it was hanging in a self-imposed sleep state - the terminator has to wake it up in order for it to be shut down properly. Added featureA & baud rates to menu. Removed feature elements from screen. Screen update tweaked - the LCD doesn't respond quicker than 100ms anyway. The time for the PPS to be shown is now dependent of screen update rate. Serial setBaud implemented. kbt_getTime() added. kbv_getYPR() added (also we also store YPR).
In main:
standby works, HOWEVER, PA0-WKUP goes into pull-down mode, and it's up to external circuitry to pull it up, which my hardware doesn't do - it's the opposite. So for the moment, I am "faking" a standby, but turning everything off, then waiting on button press. This will obviously have to change, but the hardware needs to be there for it to actually work.
BlinkerThread now exits.