Finally calculated the compensation factors for voltage calculations in the ADC.
Every writeable struct is now in KBB_types.h. I've included a version macro to make sure files are in sync.

SPIEEPROM now writes bytes across pages properly. Also, all the code for reading and writing is done in a single call, moving all spiStart, checkWIP, etc, into the write call itself.

Added eDisplay feature, now that it's actually becoming important. On LTC update, it fills in the kbb_display_t and sends it down a serial port. I need to refactor the menu code to be able to display the extra data as strings - this will allow me to choose via config to send the data on VN align, LTC align or just free-running align.

Moved __ALIGNED__ and STATIC_ASSERT to KBB_types.h, since the are the ones that use it.

Reduced the write buffers a bit to make sure we have enough SRAM for the moment. NOTE: Change it back as soon as possible!!